# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CraftType = typeof CraftType[keyof typeof CraftType];


export const CraftType = {
  smoke: 'smoke',
  brew: 'brew',
  pour: 'pour',
  vape: 'vape',
} as const;

export type CraftPhase = typeof CraftPhase[keyof typeof CraftPhase];


export const CraftPhase = {
  intro: 'intro',
  style: 'style',
  profile: 'profile',
  match: 'match',
  reveal: 'reveal',
} as const;

export type CraftBuildProfileAnswers = { [key: string]: unknown };

export interface CraftBuild {
  id: string;
  userId: string;
  venueId?: string | null;
  sessionId?: string | null;
  craft: CraftType;
  phase: CraftPhase;
  styleChoice?: string | null;
  moodChoice?: string | null;
  profileAnswers?: CraftBuildProfileAnswers;
  score?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCraftBuildRequestProfileAnswers = { [key: string]: unknown };

export interface CreateCraftBuildRequest {
  craft: CraftType;
  venueId?: string;
  sessionId?: string;
  phase?: CraftPhase;
  styleChoice?: string;
  moodChoice?: string;
  profileAnswers?: CreateCraftBuildRequestProfileAnswers;
  score?: number;
}

export type UpsertCraftBuildRequestProfileAnswers = { [key: string]: unknown };

export interface UpsertCraftBuildRequest {
  craft: CraftType;
  phase?: CraftPhase;
  styleChoice?: string;
  moodChoice?: string;
  profileAnswers?: UpsertCraftBuildRequestProfileAnswers;
  score?: number;
}

export type DesignDraftPayload = { [key: string]: unknown };

export interface DesignDraft {
  id: string;
  userId: string;
  venueId?: string | null;
  craft: CraftType;
  draftName: string;
  payload: DesignDraftPayload;
  lockedFields: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateDesignDraftRequestPayload = { [key: string]: unknown };

export interface CreateDesignDraftRequest {
  craft: CraftType;
  draftName?: string;
  payload: CreateDesignDraftRequestPayload;
  lockedFields?: string[];
}

export type UpsertDesignDraftRequestPayload = { [key: string]: unknown };

export interface UpsertDesignDraftRequest {
  craft: CraftType;
  draftName?: string;
  payload?: UpsertDesignDraftRequestPayload;
  lockedFields?: string[];
}

export interface CraftSession {
  id: string;
  userId: string;
  venueId?: string | null;
  craft: CraftType;
  buildId?: string | null;
  timerStartedAt?: string | null;
  timerDurationSecs: number;
  phase: CraftPhase;
  streakCount: number;
  lastSavedAt?: string | null;
  expiresAt?: string | null;
  createdAt: string;
}

export interface CreateCraftSessionRequest {
  craft: CraftType;
  buildId?: string;
  /**
     * @minimum 1800
     * @maximum 2400
     */
  timerDurationSecs?: number;
}

export interface UpsertCraftSessionRequest {
  craft: CraftType;
  phase?: CraftPhase;
  /** @minimum 0 */
  streakCount?: number;
  buildId?: string;
  /** @minimum 0 */
  remainingMs?: number;
}

export interface AiImageRequestBody {
  /** Craft type (pour, brew, smoke, vape) */
  craft: string;
  /** Style card identifier (e.g. smooth, spicy, light, amber) */
  styleId: string;
  /** Optional override prompt */
  prompt?: string;
}

export interface AiImageResult {
  /** Cloudinary CDN URL of the generated image */
  url: string;
  /** Whether the image was served from the cache */
  cached: boolean;
}

export type KernelModuleCraftType = typeof KernelModuleCraftType[keyof typeof KernelModuleCraftType];


export const KernelModuleCraftType = {
  smoke: 'smoke',
  pour: 'pour',
  brew: 'brew',
  vape: 'vape',
  none: 'none',
} as const;

export type KernelModuleStatus = typeof KernelModuleStatus[keyof typeof KernelModuleStatus];


export const KernelModuleStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;

export interface KernelModule {
  id: string;
  name: string;
  craftType: KernelModuleCraftType;
  slug: string;
  status: KernelModuleStatus;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  launchUrl?: string | null;
  registeredAt: string;
  updatedAt: string;
}

export type KernelModuleInputCraftType = typeof KernelModuleInputCraftType[keyof typeof KernelModuleInputCraftType];


export const KernelModuleInputCraftType = {
  smoke: 'smoke',
  pour: 'pour',
  brew: 'brew',
  vape: 'vape',
  none: 'none',
} as const;

export type KernelModuleInputStatus = typeof KernelModuleInputStatus[keyof typeof KernelModuleInputStatus];


export const KernelModuleInputStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;

export interface KernelModuleInput {
  /** @minLength 1 */
  name: string;
  craftType?: KernelModuleInputCraftType;
  /** @minLength 1 */
  slug: string;
  status?: KernelModuleInputStatus;
  description?: string;
  launchUrl?: string;
}

export type KernelModeResponseMode = typeof KernelModeResponseMode[keyof typeof KernelModeResponseMode];


export const KernelModeResponseMode = {
  sovereign: 'sovereign',
  essential: 'essential',
} as const;

export interface KernelModeResponse {
  venueId: string;
  mode: KernelModeResponseMode;
  /** @nullable */
  updatedAt?: string | null;
}

export type KernelModeInputMode = typeof KernelModeInputMode[keyof typeof KernelModeInputMode];


export const KernelModeInputMode = {
  sovereign: 'sovereign',
  essential: 'essential',
} as const;

export interface KernelModeInput {
  mode: KernelModeInputMode;
}

export type TelemetryEventInputPayload = { [key: string]: unknown };

export interface TelemetryEventInput {
  moduleId?: string;
  venueId?: string;
  /** @minLength 1 */
  eventType: string;
  payload?: TelemetryEventInputPayload;
}

export interface TelemetryDayCount {
  day: string;
  cnt: number;
}

export interface TelemetryEventTypeStat {
  event_type: string;
  cnt: number;
}

export interface TelemetryModuleUsage {
  module_name: string;
  module_slug: string;
  event_count: number;
}

export interface TelemetrySummary {
  total: number;
  dailyCounts: TelemetryDayCount[];
  topEventTypes: TelemetryEventTypeStat[];
  moduleUsage: TelemetryModuleUsage[];
  /** Ratio of build-completions to swipe-starts expressed as 0-100 */
  ritualEngagement: number;
}

export type GetCraftBuildParams = {
craft: CraftType;
};

export type GetCraftBuild200 = {
  build: CraftBuild | null;
};

export type CreateCraftBuild201 = {
  build: CraftBuild;
};

export type UpsertCraftBuild200 = {
  build: CraftBuild;
  created: boolean;
};

export type UpsertCraftBuild201 = {
  build: CraftBuild;
  created: boolean;
};

export type GetDesignDraftsParams = {
craft: CraftType;
};

export type GetDesignDrafts200 = {
  drafts: DesignDraft[];
};

export type CreateDesignDraft201 = {
  draft: DesignDraft;
};

export type UpsertDesignDraft200 = {
  draft: DesignDraft;
  created: boolean;
};

export type UpsertDesignDraft201 = {
  draft: DesignDraft;
  created: boolean;
};

export type DeleteDesignDraft200 = {
  message: string;
  id: string;
};

export type GetCraftSessionParams = {
craft: CraftType;
};

export type GetCraftSession200 = {
  session: CraftSession | null;
};

export type CreateCraftSession201 = {
  session: CraftSession;
};

export type UpsertCraftSession200 = {
  session: CraftSession;
  created: boolean;
};

export type UpsertCraftSession201 = {
  session: CraftSession;
  created: boolean;
};

export type DeleteCraftSession200 = {
  message: string;
  id: string;
};

export type ListKernelModules200 = {
  modules: KernelModule[];
};

export type RegisterKernelModule201 = {
  module: KernelModule;
};

export type IngestTelemetryEvent202 = {
  ok: boolean;
};