# lib/api-zod/src/generated/types/kernelModeInput.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KernelModeInputMode } from './kernelModeInputMode';

export interface KernelModeInput {
  mode: KernelModeInputMode;
}