# lib/api-zod/src/generated/types/kernelModeResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KernelModeResponseMode } from './kernelModeResponseMode';

export interface KernelModeResponse {
  venueId: string;
  mode: KernelModeResponseMode;
  /** @nullable */
  updatedAt?: Date | null;
}