# lib/api-zod/src/generated/types/kernelModule.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KernelModuleCraftType } from './kernelModuleCraftType';
import type { KernelModuleStatus } from './kernelModuleStatus';

export interface KernelModule {
  id: string;
  name: string;
  craftType: KernelModuleCraftType;
  slug: string;
  status: KernelModuleStatus;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  launchUrl?: string | null;
  registeredAt: Date;
  updatedAt: Date;
}