# lib/api-zod/src/generated/types/telemetryEventInput.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TelemetryEventInputPayload } from './telemetryEventInputPayload';

export interface TelemetryEventInput {
  moduleId?: string;
  venueId?: string;
  /** @minLength 1 */
  eventType: string;
  payload?: TelemetryEventInputPayload;
}